updatePresenza
non funziona da sistemare

+++ src/main/java/it/muretti/micro/repository/MurettiFreestyleMongoTemplateRepository.java
@@ -27,7 +27,7 @@ public class MurettiFreestyleMongoTemplateRepository {
 		     String nome,
 		     Date data, 
 		     String evento, 
-		     int punteggio, 
+		     double punteggio, 
 		     Date nuovaData
 		 ){
